+++ test/test_helper.rb
@@ -2,5 +2,7 @@
 
 require_relative '../lib/kanayago'
 require_relative '../lib/kanayago/cli'
+require_relative '../lib/kanayago/lsp/server'
+require_relative '../lib/kanayago/lsp/diagnostics'
 
 require 'minitest/autorun'